[dist] Vendor jquery for the client.
I want to use the CDN, but for local development and portablility, the CDN is problematic. Not everyone has an internet connection while doing dev or viewing the docs.

I'd like to figure out a way to switch from DEVELOPMENT to PRODUCTION mode when generating the static content, so that the CDN is used when docs are generated for PRODUCTION and vendored jquery is used for DEVELOPMENT. 

Until then, we have to leave stuff like this vendored.